#4 P11 plan 전체 뷰 수정, plan 개별 count 로직 추가/뷰 수정
[notice::]
progress bar는 d-flex를 사용하면 작동하지 않아서 plan-detail.html의 container, row에서 삭제함
추후 레이아웃 다시 손 봐야 함

+++ src/main/java/demo/plantodo/controller/PlanController.java
@@ -88,7 +88,7 @@ public class PlanController {
     @GetMapping("/plans")
     public String plans(Model model, HttpServletRequest request) {
         Long memberId = memberService.getMemberId(request);
-        HashMap<Plan, Float> plans = planService.findAllPlan_withCompPercent(memberId);
+        HashMap<Plan, Integer> plans = planService.findAllPlan_withCompPercent(memberId);
         model.addAttribute("plans", plans);
         return "plan/plan-list";
     }
